nba_predictor_app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.2.0
sqlalchemy>=1.4.23
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=1.8.2
mcp-core>=0.1.0
python-multipart>=0.0.5
aiohttp>=3.8.1
websockets>=10.0
pytest>=8.0.1
numpy>=1.26.4
sse-starlette>=1.6.5 